pyproject.toml
[project]
name = "pyzync"
version = "0.1.0"
description = "A tool for managing zfs filesystem backups"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dropbox>=12.0.2",
    "humanize>=4.13.0",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-exporter-otlp-proto-http>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "structlog>=25.4.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.0",
]
